mock-bank/requirements.txt
APScheduler==3.10.1

# Web Framework & Routing
Flask==3.0.0
Werkzeug==3.0.1

# Authentication & Security
PyJWT==2.8.0
Flask-Limiter==3.5.0

# Documentation (Swagger)
flasgger==0.9.7.1

# Data Generation & Simulation
Faker==19.13.0

# Production Server (for Docker)
gunicorn==21.2.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0